requirements.txt
# Core data + ML
pandas>=2.0.0
openpyxl>=3.1.0
numpy>=1.24.0
scikit-learn>=1.3.0
joblib>=1.3.0
scipy>=1.11.0

# Text + templates
beautifulsoup4>=4.12.0
lxml>=4.9.0
Jinja2>=3.1.0

# Charts
matplotlib>=3.7.0

# Web app
Flask>=3.0.0

# Config
PyYAML>=6.0

# Tests (used by CI)
pytest>=7.4.0

# Dev / CI
pytest>=7.4.0

